% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib-package.R
\docType{package}
\name{grib-package}
\alias{grib}
\alias{grib-package}
\title{grib: Read Gridded Binary ('GRIB') Files}
\description{
Reads 'GRIB' and 'GRIB2' files.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
