% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib-info.R
\name{grib2_info}
\alias{grib2_info}
\title{Read grib meta information}
\usage{
grib2_info(path, max_length = 100L)
}
\arguments{
\item{path}{File path to a gridded binary file}

\item{max_length}{The maximum number of meta records to read. Typically
this will return a single record.}
}
\description{
Read grib meta information
}
